Do the same rename as pelias folks did
