+++ test/unit/sanitizer/search.js
@@ -79,7 +79,7 @@ module.exports.tests.sanitize = function(test, common) {
         called_sanitizers.push('_categories');
         return { errors: [], warnings: [] };
       },
-      '../sanitiser/_lang': function() {
+      '../sanitizer/_lang': function() {
         called_sanitizers.push('_lang');
         return { errors: [], warnings: [] };
       }
